Move git-command to a numbered normal hook
This is possible now that the main git command is a numbered part
of the hook (6dabd17 in plushu/plushu-git, part of the
plushu/plushu#48 initiative). I mean, technically, it was possible
before as well (since *all* hooks ran before the command, due to me
not being comfortable making the command part of the hook when I
originally wrote it), but it would have been less future-proof (sort
of, in theory).

I think this was one of maybe 2 uses of early-hooks: maybe I'll remove
them altogether (plushu/plushu#46) after plushu/plushu#48 is closed
(since having all hooks be explicitly ordered means you essentially
never need to blanket-preempt your execution).